Added Backcalculated WA datasets
Added back-calculated WA dataset into DataIn\ - added html vignette to show new values.
Added redundancies to allow for the partial presence of CU, Inlet, or Enhancement data. Of note: CU, Inlet, and Stock names must be unique. To ensure this, it is recommended to add "CU" to all CU names.