Updates to Liermann_RTMB_model.R
Adding back in ADREPORT calls to allow for obj$simulate and the simulation (bootstrapping) in the overall function. This is not necessary for mcmc sampling. Likely due to uniform priors on two variables, no Std. Error's are produced.

Still working on methodical way of pulling out derived posteriors for the benchmarks.