Updates to derived_post and Liermann
derived_post.R: Fixed the estimates of the marginal mean to be of correct dimensions. There was incorrectly a call to logE_re, when the length should have been for the desired logE_tar. This was changed for logE_tar and logAlpha_tar. Spacing changes otherwise to make it easier to read.

Liermann_RTMB_model_v2.R: Added comment for running a predictions of the synoptic dataset. Removed the " - 1" from type_tar which was incorrectly indexing the watershed area regression predictions by creating a [-1] scenario of life histories. I am still concerned that this could even run without an error. Removed minor comments and spaces. Added code from running the tmbstan() call without inits and commented text for running in parallel from the tmbstan github - it does not currently work. Added marginal mean E (Srep) to the point-wise comparison plots.