Updates to current Liermann model
- Added new header to reflect Liermann and not IWAM model
- Removed comments specifying some lengths of vectors that are redundant
- Added comments to specify parameter requirements e.g. positive, zeroes
- Added parameter labelling comments
- Changed references to "imaginary" to "simulated"
- Removed second redundant setting of upper and lower limits
- Fixed comment reference of logESD and logAlphaSD as positive only for inits
- Added commented text for setting cores for parallelization

NOTE: When attempting to run mcmc sampler in parallel, it is apparent that there is one of two issues: (1) obj is not being shared between chains, or (2) some of the inits are failing when done in parallel instead of in series (this seems less likely). Concerned that this is indicative of a problem in the model that is not being caught.

- Added blank sections for upcoming prior pushforward and predictive checks and posterior predictive checks